Add command-line option to specify USB device address
Add a new command-line option, "-u", which takes a colon-separated
decimal USB bus number and device number (eg. 001:012) and ensures that
daliserver will only drive a DALI USB device if it has that address.

The intent here is that the script that starts up daliserver can read
other information from the DALI USB device (for example its serial
number) and start daliserver with options that are specific to that
device.  This will enable multiple DALI USB devices to be connected to
one host.