Multiple device support using upstart
Add an upstart configuration file for daliserver, which will be used
instead of the /etc/init.d/daliserver file when the daliserver package
is installed on a system running upstart.

This configuration file runs an instance of daliserver when a DALI USB
device is plugged in, and stops it when the device is unplugged.  A
lookup table in /etc/default/daliserver enables daliserver options to be
set based on the serial number of the device that is plugged in.