Multiple device support using upstart or systemd
Alters the Debian packaging rules to generate three separate binary
packages: daliserver, daliserver-upstart and daliserver-systemd.

The daliserver package is unchanged, i.e. it contains the
debhelper-generated sysvinit startup script.

The daliserver-upstart package has an upstart configuration file.  This
configuration file runs an instance of daliserver when a DALI USB device
is plugged in, and stops it when the device is unplugged.  A lookup
table in /etc/default/daliserver-upstart enables daliserver options to
be set based on the serial number of the device that is plugged in.

The daliserver-systemd package has a systemd unit template file and a
udev rule that runs an instance of daliserver when a DALI USB device is
plugged in, and stops it when the device is unplugged.  A lookup table
in /etc/default/daliserver-systemd enables daliserver options to be set
based on the serial number of the device that is plugged in.  A helper
script in /usr/lib/daliserver-systemd looks up the serial number, bus
number and device number based on the systemd template instance name.