Update systemd support to use built-in udev unit template support
Useful references:
 - https://github.com/systemd/systemd/issues/7109
 - https://github.com/systemd/systemd/issues/10321
 - https://github.com/systemd/systemd/commit/dcebc9bae4dcc3e844f01558c6127fc0d8745c8e

Quick summary: there are still escaping-related misfeatures in systemd
that can't be worked around using systemd-escape.  For our particular
use-case, @poettering has implemented a workaround that happens to do
what we want; this commit switches over to using it (although this
introduces a requirement that we use systemd version 236 or later).  The
general case is still broken.